+++ src/AndroidStudio/app/src/main/java/com/juankpro/ane/localnotif/LocalNotification.java
@@ -48,6 +48,8 @@ class LocalNotification {
 
     int priority = NotificationManager.IMPORTANCE_DEFAULT;
 
+    boolean showInForeground = false;
+
     String activityClassName = "";
 
     /**
@@ -81,6 +83,7 @@ class LocalNotification {
             jsonObject.putOpt("repeatInterval", repeatInterval);
             jsonObject.putOpt("activityClassName", activityClassName);
             jsonObject.putOpt("priority", priority);
+            jsonObject.putOpt("showInForeground", showInForeground);
 
             // Action data.
             for (byte anActionData : actionData) {
@@ -124,6 +127,7 @@ class LocalNotification {
             soundName = jsonObject.optString("soundName", soundName);
             activityClassName = jsonObject.optString("activityClassName", activityClassName);
             priority = jsonObject.optInt("priority", priority);
+            showInForeground = jsonObject.optBoolean("showInForeground", showInForeground);
 
             long dateTime = jsonObject.optLong("fireDate", fireDate.getTime());