+++ src/AndroidStudio/app/src/main/java/com/juankpro/ane/localnotif/LocalNotificationEventDispatcher.java
@@ -13,11 +13,11 @@ class LocalNotificationEventDispatcher {
         this.data = data;
     }
 
-    boolean dispatchInForeground() {
+    boolean dispatchWhenInForeground() {
         return dispatchWhen(ApplicationStatus.getInForeground());
     }
 
-    boolean dispatchInBackground() {
+    boolean dispatchWhenActive() {
         return dispatchWhen(ApplicationStatus.getActive());
     }